Update GET & Saving Time
Added time to redis commits for better understanding when a request was received.
Had to save the data with json string to avoid hash set in redis.

Changed retrieving of items, because it was confusing. Items get inserted in front and the parameters were used to cut of the head.